Massively simplify the Makefile...
... and avoid listing all the .a files explicitly.
Not sure if this will still work for MacOS, but it really should? Unless we can't use --static there.

For a reference setup (like the CI), a release build should now be
        make local.conf PREFIX=/install/ PEP_LIB_PATH=/install/lib PEP_INCLUDE=/install/include LIBRARY_PATH=/usr/lib/
        make -j8
        make install

For a dev build with dynamic linking (and if you don't have .a files), simply remove the --static from local.conf.